+++ examples/subscriber.rs
@@ -11,14 +11,14 @@ fn main() -> std::result::Result<(), Box<dyn std::error::Error + Send + Sync>> {
         None,
     )?)?;
 
-    host.bind_actor(
+    host.set_binding(
         "MBHRSJORBXAPRCALK6EKOBBCNAPMRTM6ODLXNLOV5TKPDMPXMTCMR4DW",
         "wascc:messaging",
         None,
         generate_config("test"),
     )?;
 
-    host.bind_actor(
+    host.set_binding(
         "MDJUPIQFWEHWE4XHPWHOJLW42SJDPVBQVDC2NV3T3O4ELXXVOXLA5M4I",
         "wascc:messaging",
         None,